+++ land_consumption/calculation.py
@@ -7,10 +7,10 @@ from land_consumption.utils import LandObjectCategory, SQM_TO_HA_FACTOR, LandUse
 def calculate_land_consumption(aoi_area: float, area_df: GeoDataFrame) -> pd.DataFrame:
     area_df['Total Land Area [ha]'] = area_df['area'] * SQM_TO_HA_FACTOR
     aoi_area_ha = aoi_area * SQM_TO_HA_FACTOR
-    area_df['% Settled Land Area'] = (area_df['Total Land Area [ha]'] / aoi_area_ha) * 100
+    area_df['% of Settled Land Area'] = (area_df['Total Land Area [ha]'] / aoi_area_ha) * 100
 
     consumed_land = area_df.apply(lambda x: x['area'] if is_land_consumed(x) else None, axis='columns')
-    area_df['% of Consumed Land Area'] = (consumed_land / consumed_land.sum() * 100).round(2)
+    area_df['% of Consumed Land Area'] = consumed_land / consumed_land.sum() * 100
 
     area_df['Land Use Object'] = area_df['category'].apply(lambda x: x.value)
     area_df['Land Use Class'] = area_df['landuse_category'].apply(lambda x: x.value)
@@ -25,9 +25,9 @@ def calculate_land_consumption(aoi_area: float, area_df: GeoDataFrame) -> pd.Dat
             'Land Use Class',
             'Total Land Area [ha]',
             '% of Consumed Land Area',
-            '% Settled Land Area',
+            '% of Settled Land Area',
         ]
-    ]
+    ].round(2)
 
 
 def is_land_consumed(feature: pd.Series) -> bool: